Change county feature
On selecting a new country via the drop down menu, the current highlight country is cleared and the newly selected country is high lighted.

Currently it used jguery to clear the dom of .leaflet-interactive[1]. I believe there is a better way using the remove() or removeLayer() function in the Leaflet library, cant get it to work. I may have initially created and added the original layer (user location layer) incorrectly.